backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-socketio==5.10.0
markdown==3.5.1
bleach==6.1.0
python-dotenv==1.0.0
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
gunicorn==21.2.0
eventlet==0.35.1
